+++ tests/testthat/test-check-params.R
@@ -1,14 +1,19 @@
 test_that("arguments are properly checked", {
   # This should fail because we can't connect to the qualtRics API
-  qualtrics_api_credentials(api_key = "1234",
-                            base_url = "yourdatacenterid.qualtrics.com")
+  qualtrics_api_credentials(
+    api_key = "1234",
+    base_url = "yourdatacenterid.qualtrics.com"
+  )
   # Query with all options
-  expect_error(getSurvey("1234",
-                         lastResponseId = "1234",
-                         startDate="2017-01-01",
-                         endDate="2017-01-31",
-                         seenUnansweredRecode = "UA",
-                         includedQuestionIds = c("QID1", "QID2"),
-                         saveDir=tempdir()),
-               "you may not have the\nrequired authorization")
+  expect_error(
+    fetch_survey("1234",
+      last_response = "1234",
+      start_date = "2017-01-01",
+      end_date = "2017-01-31",
+      unanswer_recode = "UA",
+      include_questions = c("QID1", "QID2"),
+      save_dir = tempdir()
+    ),
+    "you may not have the\nrequired authorization"
+  )
 })