+++ NEWS.md
@@ -1,9 +1,17 @@
-# qualtRics 3.0.1 [WIP]
+# qualtRics 3.1.0
 
 - New maintainer: Julia Silge
-- Simpler approach for storing API credentials as environment variables with `qualtrics_api_credentials()`
+- Add all previous contributors to DESCRIPTION as `ctb`
+- Declare testthat dependency in DESCRIPTION (reason for previous archiving from CRAN)
+- Simpler approach for storing API credentials as environment variables with `qualtrics_api_credentials()` (`registerOptions()` is now soft deprecated with a warning)
 - Simplify README (keep all existing detailed workflow documentation in vignette)
 - Relicense from GPL-3 to MIT. See [consent from authors here](https://github.com/ropensci/qualtRics/issues/95).
+- Improvements to documentation throughout
+- Renaming (with warnings on old versions) of key functions for clarity and reduction in confusion, plus improvements:
+  - `all_surveys()` (from old version of `getSurveys()`)
+  - `survey_questions()` (from old version of `getSurveyQuestions()`)
+  - `fetch_survey()` (from old version of `getSurvey()`)
+  - `read_survey()` (from old version of `readSurvey()`)
 
 
 qualtRics 3.0 (2018-02-03)